+++ src/util/mod.rs
@@ -1,3 +1,5 @@
+pub mod atomic_file;
 pub mod clock;
-#[cfg(test)]
+#[cfg(any(test, feature = "test-telemetry"))]
 pub mod test_clock;
+pub mod versioned_blob;